# backend/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Light OSS MVP API
  version: 1.0.0
servers:
  - url: http://localhost:8080
paths:
  /healthz:
    get:
      summary: Health check
      responses:
        '200':
          description: Healthy
  /api/v1/buckets:
    get:
      summary: List buckets
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Bucket list
    post:
      summary: Create bucket
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name:
                  type: string
      responses:
        '201':
          description: Bucket created
  /api/v1/buckets/{bucket}/objects:
    get:
      summary: List objects
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: bucket
          required: true
          schema:
            type: string
        - in: query
          name: prefix
          schema:
            type: string
        - in: query
          name: limit
          schema:
            type: integer
        - in: query
          name: cursor
          schema:
            type: string
      responses:
        '200':
          description: Object list
  /api/v1/buckets/{bucket}/objects/batch:
    post:
      summary: Batch upload folder files
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: bucket
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [manifest]
              properties:
                prefix:
                  type: string
                visibility:
                  type: string
                  enum: [public, private]
                manifest:
                  type: string
                  description: JSON array of {file_field, relative_path} items.
      responses:
        '201':
          description: Folder files uploaded
  /api/v1/buckets/{bucket}/folders:
    get:
      summary: List full folder tree
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: bucket
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Folder tree
    post:
      summary: Create folder
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: bucket
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [prefix, name]
              properties:
                prefix:
                  type: string
                name:
                  type: string
      responses:
        '201':
          description: Folder created
    delete:
      summary: Delete folder
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: bucket
          required: true
          schema:
            type: string
        - in: query
          name: path
          required: true
          schema:
            type: string
        - in: query
          name: recursive
          schema:
            type: boolean
          description: When true, delete the folder and all nested files and subfolders. When omitted or false, only empty folders can be deleted.
      responses:
        '204':
          description: Folder deleted
  /api/v1/buckets/{bucket}/entries:
    get:
      summary: List direct folder entries
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: bucket
          required: true
          schema:
            type: string
        - in: query
          name: prefix
          schema:
            type: string
        - in: query
          name: search
          schema:
            type: string
        - in: query
          name: limit
          schema:
            type: integer
        - in: query
          name: cursor
          schema:
            type: string
      responses:
        '200':
          description: Direct folder entries
  /api/v1/buckets/{bucket}/objects/{key}:
    get:
      summary: Download object
      responses:
        '200':
          description: Object content
    head:
      summary: Head object metadata
      responses:
        '200':
          description: Object metadata
    put:
      summary: Upload object
      security:
        - bearerAuth: []
      parameters:
        - in: header
          name: X-Object-Visibility
          schema:
            type: string
            enum: [public, private]
        - in: header
          name: X-Original-Filename
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/octet-stream:
            schema:
              type: string
              format: binary
      responses:
        '201':
          description: Object uploaded
    delete:
      summary: Soft delete object
      security:
        - bearerAuth: []
      responses:
        '204':
          description: Deleted
  /api/v1/buckets/{bucket}/objects/visibility/{key}:
    patch:
      summary: Update object visibility
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: bucket
          required: true
          schema:
            type: string
        - in: path
          name: key
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [visibility]
              properties:
                visibility:
                  type: string
                  enum: [public, private]
      responses:
        '200':
          description: Object visibility updated
  /api/v1/sign/download:
    post:
      summary: Sign download URL
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [bucket, object_key, expires_in_seconds]
              properties:
                bucket:
                  type: string
                object_key:
                  type: string
                expires_in_seconds:
                  type: integer
      responses:
        '200':
          description: Signed URL
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer